Generalize logging setup
