+++ engine/launcher/src/main/scala/org/enso/launcher/installation/DistributionUninstaller.scala
@@ -198,7 +198,7 @@ class DistributionUninstaller(
       dataRoot.toAbsolutePath.normalize
     )
     if (logsInsideData) {
-      LauncherLogging.prepareForUninstall(globalCLIOptions)
+      LauncherLogging.prepareForUninstall(globalCLIOptions.colorMode)
     }
 
     for (dirName <- knownDataDirectories) {
